requirements.txt
#dependencis
azure.functions
yfinance
azure.storage.blob
